fix: remove old buttons from docs [B1M2-5375] (#1641)
* fix: remove old button from docs [B1M2-5375]

* fix: update moon colors [B1M2-5375]

* fix: tests [B1M2-5375]
