--- next-docs/public/examples/button/Default.tsx
@@ -1,5 +1,0 @@
-import { Button } from '@heathmont/moon-components';
-
-const Example = () => <Button variant="default">Default</Button>;
-
-export default Example;
