--- next-docs/public/examples/button/Primary.tsx
@@ -1,5 +1,0 @@
-import { Button } from '@heathmont/moon-components';
-
-const Example = () => <Button variant="primary">Primary</Button>;
-
-export default Example;
