+++ next-docs/public/examples/button/Sizes.tsx
@@ -1,19 +1,11 @@
-import { Button } from '@heathmont/moon-components';
+import { Button } from '@heathmont/moon-core';
 
 const Example = () => (
   <div className="flex justify-around w-full items-center">
-    <Button variant="primary" size="xsmall">
-      Xsmall
-    </Button>
-    <Button variant="primary" size="small">
-      Small
-    </Button>
-    <Button variant="primary" size="medium">
-      Medium
-    </Button>
-    <Button variant="primary" size="large">
-      Large
-    </Button>
+    <Button size="small">Small</Button>
+    <Button>Medium is default</Button>
+    <Button size="large">Large</Button>
+    <Button size="xlarge">xLarge</Button>
   </div>
 );
