--- next-docs/public/examples/button/Tertiary.tsx
@@ -1,5 +1,0 @@
-import { Button } from '@heathmont/moon-components';
-
-const Example = () => <Button variant="tertiary">Tertiary</Button>;
-
-export default Example;
