+++ next-docs/public/examples/button/__tests__/index.test.tsx
@@ -1,41 +1,34 @@
 /**
  * @jest-environment jsdom
  */
-
 import React from 'react';
 import { moonDesignLight, ThemeProvider } from '@heathmont/moon-themes';
 import renderer from 'react-test-renderer';
-import Primary from '../Primary';
-import Secondary from '../Secondary';
-import Tertiary from '../Tertiary';
-import Sizes from '../Sizes';
 import Animations from '../Animations';
+import Disabled from '../Disabled';
 import Icons from '../Icons';
-import IsIcon from '../IsIcon';
+import Sizes from '../Sizes';
+import Variants from '../Variants';
 import FullWidth from '../FullWidth';
 
 const renderWithTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
 );
 
+const renderWithThemeRTL = (component: JSX.Element) => (
+  <ThemeProvider theme={moonDesignLight}>
+    <div dir="rtl">{component}</div>
+  </ThemeProvider>
+);
+
 describe('Button', () => {
-  it('renders Primary variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Primary />));
+  it('renders different variants', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Variants />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders Secondary variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Secondary />));
-    expect(testRenderer.toJSON()).toMatchSnapshot();
-  });
-
-  it('renders Tertiary variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Tertiary />));
-    expect(testRenderer.toJSON()).toMatchSnapshot();
-  });
-
-  it('renders Default variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Primary />));
+  it('renders different variants RTL', () => {
+    const testRenderer = renderer.create(renderWithThemeRTL(<Variants />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
@@ -44,23 +37,23 @@ describe('Button', () => {
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders different animations', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Animations />));
-    expect(testRenderer.toJSON()).toMatchSnapshot();
-  });
-
-  it('renders leading and trailing icons', () => {
+  it('renders different icons', () => {
     const testRenderer = renderer.create(renderWithTheme(<Icons />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders rounded "isIcon" variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<IsIcon />));
+  it('renders Full width property', () => {
+    const testRenderer = renderer.create(renderWithTheme(<FullWidth />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders "Full width" variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<FullWidth />));
+  it('renders disabled state', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Disabled />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders different animations', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Animations />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 });
