--- next-docs/public/examples/buttonNew/Icons.tsx
@@ -1,23 +1,0 @@
-import { Button } from '@heathmont/moon-core';
-import { GenericSettings } from '@heathmont/moon-icons';
-
-const Example = () => (
-  <div className="flex justify-around w-full items-center">
-    <Button iconLeft={<GenericSettings />}>IconLeft</Button>
-    <Button iconLeft>
-      <GenericSettings />
-      IconLeft
-    </Button>
-    <Button iconRight={<GenericSettings />}>IconRight</Button>
-    <Button iconRight>
-      IconRight
-      <GenericSettings />
-    </Button>
-    <Button iconOnly={<GenericSettings />} />
-    <Button iconOnly>
-      <GenericSettings />
-    </Button>
-  </div>
-);
-
-export default Example;
